\name{accessorsFLBRP}
\docType{methods}
\alias{model,FLBRP-method}
\alias{model<-,FLBRP,formula-method}
\alias{params,FLBRP-method}
\alias{params<-,FLBRP,FLPar-method}
\alias{refpts,FLBRP-method}
\alias{refpts<-,FLBRP,refpts-method}
\alias{fbar,FLBRP-method}
\alias{fbar<-,FLBRP,FLQuant-method}
\alias{fbar.obs,FLBRP-method}
\alias{fbar.obs<-,FLBRP,FLQuant-method}
\alias{landings.obs,FLBRP-method}
\alias{landings.obs<-,FLBRP,FLQuant-method}
\alias{discards.obs,FLBRP-method}
\alias{discards.obs<-,FLBRP,FLQuant-method}
\alias{rec.obs,FLBRP-method}
\alias{rec.obs<-,FLBRP,FLQuant-method}
\alias{ssb.obs,FLBRP-method}
\alias{ssb.obs<-,FLBRP,FLQuant-method}
\alias{profit.obs,FLBRP-method}
\alias{profit.obs<-,FLBRP,FLQuant-method}
\alias{landings.sel,FLBRP-method}
\alias{landings.sel<-,FLBRP,FLQuant-method}
\alias{discards.sel,FLBRP-method}
\alias{discards.sel<-,FLBRP,FLQuant-method}
\alias{bycatch.harvest,FLBRP-method}
\alias{bycatch.harvest<-,FLBRP,FLQuant-method}
\alias{stock.wt,FLBRP-method}
\alias{stock.wt<-,FLBRP,FLQuant-method}
\alias{landings.wt,FLBRP-method}
\alias{landings.wt<-,FLBRP,FLQuant-method}
\alias{discards.wt,FLBRP-method}
\alias{discards.wt<-,FLBRP,FLQuant-method}
\alias{bycatch.wt,FLBRP-method}
\alias{bycatch.wt<-,FLBRP,FLQuant-method}
\alias{m,FLBRP-method}
\alias{m<-,FLBRP,FLQuant-method}
\alias{mat,FLBRP-method}
\alias{mat<-,FLBRP,FLQuant-method}
\alias{harvest.spwn,FLBRP-method}
\alias{harvest.spwn<-,FLBRP,FLQuant-method}
\alias{m.spwn,FLBRP-method}
\alias{m.spwn<-,FLBRP,FLQuant-method}
\alias{availability,FLBRP-method}
\alias{availability<-,FLBRP,FLQuant-method}
\alias{price,FLBRP-method}
\alias{price<-,FLBRP,FLQuant-method}
\alias{vcost,FLBRP-method}
\alias{vcost<-,FLBRP,FLQuant-method}
\alias{fcost,FLBRP-method}
\alias{fcost<-,FLBRP,FLQuant-method}
\title{Accessor method for FLBRP and refpts slots}
\description{
}
\section{Generic function}{
methodName(class1, class2) 
}
\section{Methods}{\describe{
\item{signature(obj1=class1,obj2=class2):}{}
}}
\author{The FLR Team}
\seealso{\link[originalPkg]{methodName}}
\examples{
}
\keyword{methods}
