\name{brp}
\docType{methods}
\alias{brp}
\title{brp method for estimating BRPs}
\description{
  Estimates all specified reference points and returns an object of class \code{\link{FLBP-class}}.
}
\section{Generic function}{
 brp(object) 
}
\section{Methods}{\describe{
\item{}{}
  }
}
\author{The FLR Team}
\seealso{}
\examples{
}
\keyword{methods}
