\name{}
\docType{class}
% \alias{className-class}
% \alias{className}
% \alias{className-methods}
% \alias{className, inputClass-method}
\alias{refpts<-,FLBRP,refpts-method}
\alias{refpts<-,FLBRP,numeric-method}
\alias{refpts,FLBRP-method}
\alias{propagate,FLBRP-method}
\alias{iter,FLBRP-method}
\alias{FLBRP,missing,missing-method}
\alias{FLBRP,missing,FLSR-method}
\alias{FLBRP,FLStock,FLSR-method}
\alias{FLBRP,FLStock,missing-method}

\title{}
\description{
}
\section{Slots}{\describe{
\item{}{. \code{classOfSlot}.}
\item{}{. \code{character}.}
}}
\section{Extends}{\describe{
}}

\section{Constructor}{
\describe{
}}
\section{Constructor Arguments}{
  \describe{
  \item{object}{An input object.}
  }
}
\section{Methods}{\describe{
\item{\link[pkg]{method}(pkg):}{}
\item{}{\code{signature(x=classsName)}}
}}
\author{The FLR Team}
\seealso{}
\examples{
}
\keyword{classes}
