\name{}
\docType{methods}
% \alias{methodName,class1,class2-method}
\alias{landings.n,FLBRP-method}
\alias{discards.n,FLBRP-method}
\alias{stock.n,FLBRP-method}
\alias{catch.n,FLBRP-method}
\alias{catch.wt,FLBRP-method}
\alias{catch.sel,FLBRP-method}
\alias{catch.obs,FLBRP-method}
\alias{yield.obs,FLBRP-method}
\alias{rec,FLBRP-method}
\alias{rec.hat,FLBRP-method}
\alias{harvest,FLBRP,missing-method}
\alias{yield,FLBRP-method}
\alias{yield.hat,FLBRP-method}
\alias{discards,FLBRP-method}
\alias{discards.hat,FLBRP-method}
\alias{landings,FLBRP-method}
\alias{landings.hat,FLBRP-method}
\alias{stock,FLBRP-method}
\alias{stock.hat,FLBRP-method}
\alias{ssb,FLBRP-method}
\alias{ssb.hat,FLBRP-method}
\alias{catch,FLBRP-method}
\alias{catch.hat,FLBRP-method}
\alias{}
\title{}
\description{
}
\section{Generic function}{
methodName(class1, class2) 
}
\section{Methods}{\describe{
\item{signature(obj1=class1,obj2=class2):}{}
}}
\author{The FLR Team}
\seealso{\link[originalPkg]{methodName}}
\examples{
}
\keyword{methods}
