\name{kobe}
\docType{methods}
\alias{kobe,FLQuants,FLQuants-method}
\alias{kobe,FLQuants,FLQuant-method}
\alias{kobe,FLQuant,FLQuants-method}
\alias{kobe,FLQuant,FLQuant-method}
\alias{kobe,FLQuants,missing-method}
\alias{kobe,array,missing-method}
\alias{kobe,data.frame,missing-method}
\title{Kobe plots of stock status against MSY}
\description{
}
\section{Generic function}{
% methodName(object) 
}
\section{Methods}{\describe{
\item{signature(object=oneClass) :}{}
}}
\author{The FLR Team}
\seealso{}
\examples{
}
\keyword{methods}
