\name{msy}
\docType{methods}
\alias{msy}
\alias{msy-methods}
\alias{msy,FLBRP-method}
\title{Maximum Sustainable Yield}
\description{
}
\section{Generic function}{
% methodName(object) 
}
\section{Methods}{\describe{
\item{signature(object=oneClass) :}{}
}}
\author{The FLR Team}
\seealso{}
\examples{
}
\keyword{methods}
