\name{plotFLBRP}
\docType{methods}
\alias{plot,FLBRP,missing-method}
\alias{plot,FLBRP,FLStock-method}
\title{plot methods for FLBRP}
\description{
}
\section{Generic function}{
methodName(class1, class2) 
}
\section{Methods}{\describe{
\item{signature(obj1=class1,obj2=class2):}{}
}}
\author{The FLR Team}
\seealso{\link[originalPkg]{methodName}}
\examples{
}
\keyword{methods}
