\name{}
\docType{class}
% \alias{className-class}
% \alias{className}
% \alias{className-methods}
% \alias{className, inputClass-method}
\alias{refpts}
\alias{refpts-methods}
\alias{refpts,array-method}
\alias{refpts,missing-method}
\alias{refpts,numeric-method}
\alias{refpts,logical-method}
\alias{refpts,refpts-method}
\alias{show,refpts-method}
\alias{propagate,refpts-method}
\title{}
\description{
}
\section{Slots}{\describe{
\item{}{. \code{classOfSlot}.}
\item{}{. \code{character}.}
}}
\section{Extends}{\describe{
}}

\section{Constructor}{
\describe{
}}
\section{Constructor Arguments}{
  \describe{
  \item{object}{An input object.}
  }
}
\section{Methods}{\describe{
\item{\link[pkg]{method}(pkg):}{}
\item{}{\code{signature(x=classsName)}}
}}
\author{The FLR Team}
\seealso{}
\examples{
}
\keyword{classes}
