\name{setPlusGroups}
\docType{methods}
\alias{setPlusGroup,FLBRP,numeric-method}
\title{Set plusgroup for FLBRP}
\description{
}
\section{Generic function}{
 setPlusGroups(x, plusgroup, ...) 
}
\section{Methods}{\describe{
\item{}{}
  }
}
\author{The FLR Team}
\seealso{}
\examples{
}
\keyword{methods}
