\name{}
\docType{methods}
%\alias{methodName}
%\alias{methodName-methods}
%\alias{methodName,class1-method}
\alias{spr}
\alias{spr-methods}
\alias{spr,FLBRP-method}
\alias{ypr}
\alias{ypr-methods}
\alias{ypr,FLBRP-method}
\alias{spr0}
\alias{spr0-methods}
\alias{spr0,FLBRP,missing,missing-method}

\alias{}
\title{}
\description{
}
\section{Generic function}{
% methodName(object) 
}
\section{Methods}{\describe{
\item{signature(object=oneClass) :}{}
}}
\author{The FLR Team}
\seealso{}
\examples{
}
\keyword{methods}
