\name{}
\docType{methods}
\alias{brp}
\alias{brp-methods}
\alias{brp,FLBRP-method}
\alias{computeRefpts}
\alias{computeRefpts-methods}
\alias{computeRefpts,FLBRP-method}
\title{}
\description{
}
\section{Generic function}{
% methodName(object) 
}
\section{Methods}{\describe{
\item{signature(object=oneClass) :}{}
}}
\author{The FLR Team}
\seealso{}
\examples{
}
\keyword{methods}
