\name{}
\docType{methods}
%\alias{methodName}
%\alias{methodName-methods}
%\alias{methodName,class1-method}
\alias{computeRefpts,FLBRP-method}
\alias{brp,FLBRP-method}
\alias{}
\title{}
\description{
}
\section{Generic function}{
% methodName(object) 
}
\section{Methods}{\describe{
\item{signature(object=oneClass) :}{}
}}
\author{The FLR Team}
\seealso{}
\examples{
}
\keyword{methods}
