\name{hcrYield}
\docType{methods}
\alias{computeFbar,FLBRP-method}
\alias{hcrYield,FLBRP,FLQuant-method}
\alias{hcrYield,FLBRP,numeric-method}
\title{Calculate yield and fishing mortalities from FLBRP}
\description{
}
\section{Generic function}{
% methodName(object) 
}
\section{Methods}{\describe{
\item{signature(object=oneClass) :}{}
}}
\author{The FLR Team}
\seealso{}
\examples{
}
\keyword{methods}
