\name{}
\docType{methods}
%\alias{methodName}
%\alias{methodName-methods}
%\alias{methodName,class1-method}
\alias{kobe,FLQuants,FLQuants-method}
\alias{kobe,FLQuants,FLQuant-method}
\alias{kobe,FLQuant,FLQuants-method}
\alias{kobe,FLQuant,FLQuant-method}
\alias{kobe,FLQuants,missing-method}
\alias{kobe,array,missing-method}
\alias{kobe,data.frame,missing-method}
\alias{}
\title{}
\description{
}
\section{Generic function}{
% methodName(object) 
}
\section{Methods}{\describe{
\item{signature(object=oneClass) :}{}
}}
\author{The FLR Team}
\seealso{}
\examples{
}
\keyword{methods}
